some little fixings
