+++ js/main.js
@@ -1,2 +1,2 @@
 import './modules/data.js';
-
+import './modules/offer.js';
