+++ js/modules/data.js
@@ -35,7 +35,7 @@ const PHOTOS = [
   'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
 ];
 
-const NUMBER_OF_COPIES = 10;
+const NUMBER_OF_COPIES = 2;
 
 const createAuthor = (index) => {
   if (index < 9) {
@@ -79,7 +79,7 @@ const getCreateCard = (index) => (
   }
 );
 
-const createCards = new Array(NUMBER_OF_COPIES).fill('').map((value, index) =>
+const newOffers = new Array(NUMBER_OF_COPIES).fill('').map((value, index) =>
   getCreateCard(index));
 
-export {createCards};
+export {newOffers, PHOTOS, FEATURES};
